👽️ 因應粉圓 release 改為上傳 ttf 變更流程

+++ for_me.sh
@@ -14,35 +14,46 @@ if [ ! -d "$HOME/.fonts" ] ; then
 fi
 
 url=$(curl -s https://api.github.com/repos/justfont/open-huninn-font/releases/latest | grep "browser_download_url" | awk '{ print $2 }' | sed 's/"//g')
-zipname=$(basename $url .zip)
+extension="${url##*.}"
 
-echo "Now download $zipname from $url"
-
-# Let's download font
-# download 粉圓
-wget -O huninn.zip $url
-
-# unzip
-unzip huninn.zip
-
-# check zip file download and upzip succefully
-if [ ! -d "$zipname" ] ; then
-	echo -e "\n\n"
-	echo -e "\a========= The font file went wrong! Please try again later. =========\n"
-	exit 1
+if [ "$extension" = "zip" ]; then
+    zipname=$(basename $url .zip)
+    
+    echo "Now download $zipname from $url"
+    
+    # Let's download font
+    # download 粉圓
+    wget -O huninn.zip $url
+    
+    # unzip
+    unzip huninn.zip
+    
+    # check zip file download and upzip succefully
+    if [ ! -d "$zipname" ] ; then
+        echo -e "\n\n"
+        echo -e "\a========= The font file went wrong! Please try again later. =========\n"
+        exit 1
+    fi
+    
+    # copy fonts to ~/.fonts
+    echo -e "\n\n"
+    cp -i $zipname/*.ttf $HOME/.fonts
+    
+    # remove zip and unzip dir
+    rm huninn.zip
+    rm -r $zipname
+else
+    filename=$(basename $url)
+    echo "Now download $filename from $url"
+    wget $url
+    # copy fonts to ~/.fonts
+    echo -e "\n\n"
+    mv *.ttf $HOME/.fonts
 fi
-
-# copy fonts to ~/.fonts
-echo -e "\n\n"
-cp -i $zipname/*.ttf $HOME/.fonts
 
 # refresh fonts cache
 echo -e "\n"
 sudo fc-cache -fv
-
-# remove zip and unzip dir
-rm huninn.zip
-rm -r $zipname
 
 # finally
 echo -e "\n\n"
